# performance improvement on certain operation