--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/yl2chen/cidranger
 
 go 1.13
 
-require github.com/stretchr/testify v1.4.0
+require (
+	github.com/stretchr/testify v1.6.1
+	gopkg.in/yaml.v2 v2.2.2 // indirect
+)