@dotwaffle 2020-12-26
LGTM! All the tests in my code pass when replacing with this modified version, looks like an easy win for double the performance here!