@panga 2021-04-01
This PR improves insert performance by 60% for my use case and also reduces heap memory usage by 25%. Can we get this merged and released?